Add files via upload
This Currency_Converter project using PHP,HTML,MySQL database,JavaScript and Tomcat server is used . This project doesn't consist of any API for dynamic currency values but uses the SQL database for the previously stored values in it. So, it is static. The provided CSV file consists of previously stored currency ratio rates for each conversion of one currency to other. I have provided "dbcc creation.php","creationtablecc.php","insertioncc.php" for creating Database,Table and to insert values as in CSV. "cleardataDB.php" is an optional php file to clear the table values. (or) Manually creation of database,table  & insertion,updation etc can be done in the server repository.